Separated executionManager_test into executionManager_test and matchingPersistence_test
